backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
gunicorn==21.2.0
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL
pandas==2.1.3
openpyxl==3.1.2
APScheduler==3.10.4
twilio==8.10.0
sendgrid==6.11.0
python-dateutil==2.8.2
marshmallow==3.20.1
